refactor(plines): move horizontal size functions to plines.c
fix style.
