+++ src/nvim/getchar.c
@@ -34,6 +34,7 @@
 #include "nvim/memory.h"
 #include "nvim/message.h"
 #include "nvim/misc1.h"
+#include "nvim/plines.h"
 #include "nvim/keymap.h"
 #include "nvim/garray.h"
 #include "nvim/move.h"
